requirements.txt
discord.py==2.2.2
discord-py-interactions
discord-py-slash-command
openai